Move grid-mapping-specific content out of common load_cube__activate code.
