+++ lib/iris/tests/unit/fileformats/netcdf/load_cube/load_cube__activate/test__additional.py
@@ -17,12 +17,12 @@ Both of those supply an "engine" with an "activate" method
 import iris.tests as tests
 
 
-from iris.tests.unit.fileformats.netcdf.load_cube.load_cube__activate import (
-    Mixin__nc_load_actions,
+from iris.tests.unit.fileformats.netcdf.load_cube.load_cube__activate.test__grid_mappings import (
+    Mixin__grid_mapping,
 )
 
 
-class Test__additional(Mixin__nc_load_actions, tests.IrisTest):
+class Test__additional(Mixin__grid_mapping, tests.IrisTest):
     # Run grid-mapping tests with non-Pyke (actions)
     use_pyke = False
